refactor(#3281): extract indexed builtin ctor dispatch from compileNewExpression (WAVE-C slice 2)
Lift the indexed built-in constructor dispatch band (~630 LOC) VERBATIM out of
`compileNewExpression` (src/codegen/expressions/new-super.ts) into a new sibling
module `src/codegen/expressions/new-indexed.ts` (`tryCompileIndexedBuiltinNew`).

Covered arms: ArrayBuffer (including resizable ArrayBuffer), DataView, Array.

Byte-identity sentinel lift: each arm's internal `return`s are left verbatim; a
single `NEW_INDEXED_FALLTHROUGH` sentinel signals "not one of these indexed
builtins" so the caller resumes to its final unsupported-ctor `reportError`. The
band-shared helper `inferArrayElementType` is now `export`ed from new-super.ts.

- compileNewExpression: ~1,994 -> 1,357 LOC (now under the 1,500 target)
- new-indexed.ts: 685 LOC (< 1,500 threshold)
- prove-emit-identity: IDENTICAL across all 39 gc/standalone/wasi emits
- tsc --noEmit clean; tests/issue-3281.test.ts (15 cases) green

Slice 2 of 2 (#3281, subtask of #3182) — completes the compileNewExpression
decomposition. Stacked on slice 1 (built-in globals). Issue -> status: done.

✓
